+++ src/bvtest/repo_helper.go
@@ -1,4 +1,4 @@
-package test
+package bvtest
 
 import (
 	"fmt"
